[modularize] Create `TasksCache`, move `publish_actor_class_to_key` out of `actor` module.
